requirements.txt
Appium-Python-Client==0.24
behave==1.2.5
coverage==4.2
enum34==1.1.6
funcsigs==1.0.2
mock==2.0.0
nose==1.3.7
parse==1.8.2
parse-type==0.3.4
pbr==3.1.1
selenium==3.3.1
six==1.10.0
